allow SpawnDebugger after user close
Previously we would require the user to call KillDebugger() in order to be able
to call SpawnDebugger() again, even when the user had manually closed the
debugger. Fix that by checking whether a previously-spawned debugger process
is still running when the user calls SpawnDebugger again. If it is, print out
the process information (since it can happen that the debugger becomes a zombie
or something and the user needs more information in order to kill it). If it
isn't, spawn a new debugger process as requested.

Factor out the debugger process info cleanup logic, since it's being used in 5
places. Would probably be better to use a smart pointer, but for now this is
clear enough.

Ensure that every branch that returns `false` in `SpawnDebugger` and
`KillDebugger` prints a message explaining what happened.

Fixes #483.