Add debugger menu item to open settings dir.
Sometimes the user needs to get to their settings folder (e.g. to inspect,
share, or delete it), but it's buried deep somewhere they never go on their hard
drive. Add a menu item that opens the File Explorer at the location where the
settings are located.

Ideally users would never need this, but weird things happen and it's nice to
empower users to take care of it when they do.

Fixes #481.