Fix debugger warnings in unit tests
The sml client Agent destructor calls `KillDebugger` as a precautionary cleanup,
though the debugger process may never have been spawned. This lead to hundreds
of warnings in the unit tests, where agents are constantly created and
destroyed. Users are likely seeing this message in their applications as well.

Add an `ignoreNonExistent` parameter to `KillDebugger` that prevents a warning
from being printed when the debugger process is not present. Don't expose to
SWIG bindings for now, but use it in the Agent destructor, which takes care of
spurious warnings.

Thanks to Aaron Mininger for reporting.