Safeley handle nulls in WME constructors from SML
If a client passes null to any of the `Create*WME` constructors, Soar segfaults
due to accessing invalid memory. This is because it assumes non-null strings for
attribute names and values.

Fix this by checking for null and using `"nil"` when required. This is in-line
with Soar behavior elsewhere (where the `nil` Symbol gets stringified into an
attribute name). Also print a warning so that the user knows that this happened.
It was likely an accident.

The fixes work for all of our SWIG-bindings, so we only need to test in one. Use
Python to test the conversions to `"nil"`.

Fixes #485.